+++ components/afArrayField/afArrayField.js
@@ -5,7 +5,7 @@ Template.afArrayField.helpers({
     return AutoForm.getTemplateName('afArrayField', this.template, this.name);
   },
   innerContext: function afArrayFieldContext() {
-    var c = AutoForm.Utility.normalizeContext(this, "afArrayField");
+    var c = AutoForm.Utility.getComponentContext(this, "afArrayField");
     var name = c.atts.name;
     var fieldMinCount = c.atts.minCount || 0;
     var fieldMaxCount = c.atts.maxCount || Infinity;