+++ components/afQuickField/afQuickField.js
@@ -2,13 +2,13 @@
 
 Template.afQuickField.helpers({
   isGroup: function afQuickFieldIsGroup() {
-    var c = AutoForm.Utility.normalizeContext(this, "afQuickField");
+    var c = AutoForm.Utility.getComponentContext(this, "afQuickField");
     // Render a group of fields if we expect an Object and we don't have options
     // and we have not overridden the type
     return (c.defs.type === Object && !c.atts.options && !c.atts.type);
   },
   isFieldArray: function afQuickFieldIsFieldArray() {
-    var c = AutoForm.Utility.normalizeContext(this, "afQuickField");
+    var c = AutoForm.Utility.getComponentContext(this, "afQuickField");
     // Render an array of fields if we expect an Array and we don't have options
     // and we have not overridden the type
     return (c.defs.type === Array && !c.atts.options && !c.atts.type);
@@ -19,7 +19,7 @@ Template.afQuickField.helpers({
     return _.omit(this, 'fields', 'omitFields');
   },
   isHiddenInput: function afQuickFieldIsHiddenInput() {
-    var c = AutoForm.Utility.normalizeContext(this, "afQuickField");
+    var c = AutoForm.Utility.getComponentContext(this, "afQuickField");
     var inputType = c.atts.type;
     if (inputType) {
       var componentDef = AutoForm._inputTypeDefinitions[inputType];