+++ components/afQuickFields/afQuickFields.js
@@ -1,12 +1,23 @@
+/* global AutoForm */
+
 Template.afQuickFields.helpers({
   quickFieldAtts: function afQuickFieldsQuickFieldAtts() {
+    var afQuickFieldsComponentAtts, defaultOptions, atts = {};
+
     // Get the attributes that were on the afQuickFields component
-    var afQuickFieldsComponentAtts = Template.parentData(1);
+    afQuickFieldsComponentAtts = Template.parentData(1);
     // It's possible to call {{> afQuickFields}} with no attributes, in which case we
     // don't want the "attributes" because they're really just the parent context.
     if (afQuickFieldsComponentAtts.atts) {
       afQuickFieldsComponentAtts = {};
     }
-    return _.extend({options: "auto"}, afQuickFieldsComponentAtts, this);
+
+    // Add default options from schema/allowed
+    defaultOptions = AutoForm._getOptionsForField(this.name);
+    if (defaultOptions) {
+      atts.options = defaultOptions;
+    }
+
+    return _.extend(atts, afQuickFieldsComponentAtts, this);
   }
 });